Warrior: Move aura focused rage to focused rage
Mortal strike is already overloaded. In PVP you want to know if a target has the mortal strike debuff. In PVE you want to see Shattered Defenses, however that can be shown on execute aswell. Ideally you want to see the stacks of focused rage on the button itself, so you don't press it when you are at 3 stacks. This also helps if you have rage between pulls and want to stack a few stacks already.